+++ parsley/test/Parsley/Alternative/Test.hs
@@ -6,7 +6,6 @@ import TestUtils
 import qualified Parsley.Alternative.Parsers as Parsers
 
 import Prelude hiding ()
-import Parsley (runParser)
 
 tests :: TestTree
 tests = testGroup "Alternative" [ coproductTests