+++ parsley/test/Parsley/Applicative/Test.hs
@@ -6,7 +6,6 @@ import TestUtils
 import qualified Parsley.Applicative.Parsers as Parsers
 
 import Prelude hiding ()
-import Parsley (runParser)
 import Parsley.Applicative (unit)
 
 tests :: TestTree