+++ parsley/test/Parsley/Register/Test.hs
@@ -6,7 +6,6 @@ import TestUtils
 import qualified Parsley.Register.Parsers as Parsers
 
 import Prelude hiding ()
-import Parsley (runParser)
 
 -- NewRegister is required for any test and can only be tested with another component
 -- so it is omitted